+++ crates/protos/src/lib.rs
@@ -5,3 +5,7 @@
 pub mod protos;
 
 pub use protos::*;
+
+/// The descriptor set used to generate the SDK's bundled protobuf messages.
+#[cfg(feature = "descriptors")]
+pub const FILE_DESCRIPTOR_SET: &[u8] = include_bytes!(concat!(env!("OUT_DIR"), "/descriptors.bin"));